Change to from parsing temperature as int to a float.

+++ src/main/resources/static/javascript/graphGenerator.js
@@ -87,7 +87,7 @@ function changeTemperatureUnit(unit) {
 
         // Change in unit
         if (currentTempUnitText === "°F" && !isNaN(Number(currentTempReading.innerText))) {
-            currentTempReading.innerText = convertFahrenheitToCelsius(parseInt(currentTempReading.innerText));
+            currentTempReading.innerText = convertFahrenheitToCelsius(parseFloat(currentTempReading.innerText));
         }
         currentTempUnit.innerText = "°C";
 
@@ -98,7 +98,7 @@ function changeTemperatureUnit(unit) {
 
         // Change in unit
         if (currentTempUnitText === "°C" && !isNaN(Number(currentTempReading.innerText))) {
-            currentTempReading.innerText = convertCelsiusToFahrenheit(parseInt(currentTempReading.innerText));
+            currentTempReading.innerText = convertCelsiusToFahrenheit(parseFloat(currentTempReading.innerText));
         }
         currentTempUnit.innerText = "°F";
